refactor: expose CLI main function externally (temp)
